Refactoring the signup__free spec into an onboarding flow spec.
This commit renames signup__free to heroflow__write and removes the new site sign up flow which is out of scope for onboarding.
